requirements.txt
g4f
setuptools
wheel
requests==2.31.0
moviepy==1.0.3
termcolor==2.4.0
flask==3.0.0
curl_cffi
flask-cors==4.0.0
playsound==1.3.0
Pillow==9.5.0
python-dotenv==1.0.0
srt_equalizer==0.1.8
platformdirs==4.1.0
undetected_chromedriver
assemblyai
brotli
google-api-python-client 
oauth2client
openai
google-generativeai
PyYAML==6.0.1
